# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Citycard",
        "version": "1.0.0"
    },
    "paths": {
        "/api/cards": {
            "get": {
                "tags": [
                    "Card"
                ],
                "summary": "Get all authenticated user's cards",
                "operationId": "e35184e8a68a0379ee92df6444996c8a",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "user_id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "number": {
                                                        "type": "number",
                                                        "example": "12345678909"
                                                    },
                                                    "type": {
                                                        "type": "string",
                                                        "example": "preferential"
                                                    },
                                                    "current_balance": {
                                                        "type": "number",
                                                        "example": "104.5"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/cards/{cardId}": {
            "get": {
                "tags": [
                    "Card"
                ],
                "summary": "Get user's card by cardId",
                "operationId": "288b740244f2846c13fcfede9c08e0cb",
                "parameters": [
                    {
                        "name": "cardId",
                        "in": "path",
                        "description": "ID of card to return",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "user_id": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "number": {
                                            "type": "number",
                                            "example": "12345678909"
                                        },
                                        "type": {
                                            "type": "string",
                                            "example": "Спеціальний"
                                        },
                                        "current_balance": {
                                            "type": "number",
                                            "example": "104.5"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}